+++ mate-panel/panel-test-applets.c
@@ -178,7 +178,7 @@ static void
 load_applet_from_command_line (void)
 {
 	guint size = 24, orient = PANEL_ORIENTATION_TOP;
-	gint i;
+	gsize i;
 
 	g_assert (cli_iid != NULL);
